# 2.瑞吉外卖/第二阶段优化/资料/day03/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "172.17.0.60:8200",
  "basePath": "/",
...
  "paths": {
...
    "/order": {
      "post": {
        "tags": [
          "订单"
        ],
        "summary": "新增",
        "operationId": "saveUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orders",
            "description": "orders",
            "required": true,
            "schema": {
              "$ref": "#/definitions/订单表1",
              "originalRef": "订单表1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«订单表»",
              "originalRef": "返回值«订单表»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "订单"
        ],
        "summary": "修改",
        "operationId": "updateUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orders",
            "description": "orders",
            "required": true,
            "schema": {
              "$ref": "#/definitions/订单表1",
              "originalRef": "订单表1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«订单表»",
              "originalRef": "返回值«订单表»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "订单"
        ],
        "summary": "删除",
        "operationId": "deleteUsingDELETE_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值",
              "originalRef": "返回值"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/order/page": {
      "get": {
        "tags": [
          "订单"
        ],
        "summary": "分页",
        "operationId": "pageUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "地址",
            "required": false,
            "type": "string"
          },
          {
            "name": "addressBookId",
            "in": "query",
            "description": "地址",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "amount",
            "in": "query",
            "description": "实收金额",
            "required": false,
            "type": "number"
          },
          {
            "name": "beginTime",
            "in": "query",
            "description": "beginTime",
            "required": false,
            "type": "string"
          },
          {
            "name": "checkoutTime",
            "in": "query",
            "description": "结账时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "consignee",
            "in": "query",
            "description": "收货人",
            "required": false,
            "type": "string"
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "endTime",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "number",
            "in": "query",
            "description": "订单号",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderTime",
            "in": "query",
            "description": "下单时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "payMethod",
            "in": "query",
            "description": "支付方式 1微信，2支付宝",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "remark",
            "in": "query",
            "description": "备注",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "下单用户",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userName",
            "in": "query",
            "description": "用户名",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«Page«订单表»»",
              "originalRef": "返回值«Page«订单表»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/submit": {
      "post": {
        "tags": [
          "订单"
        ],
        "summary": "提交订单",
        "operationId": "submitUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orders",
            "description": "orders",
            "required": true,
            "schema": {
              "$ref": "#/definitions/订单表1",
              "originalRef": "订单表1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«订单表»",
              "originalRef": "返回值«订单表»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/userPage": {
      "get": {
        "tags": [
          "订单"
        ],
        "summary": "用户订单分页",
        "operationId": "pageUsingGET_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "地址",
            "required": false,
            "type": "string"
          },
          {
            "name": "addressBookId",
            "in": "query",
            "description": "地址",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "amount",
            "in": "query",
            "description": "实收金额",
            "required": false,
            "type": "number"
          },
          {
            "name": "checkoutTime",
            "in": "query",
            "description": "结账时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "consignee",
            "in": "query",
            "description": "收货人",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "number",
            "in": "query",
            "description": "订单号",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderTime",
            "in": "query",
            "description": "下单时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "payMethod",
            "in": "query",
            "description": "支付方式 1微信，2支付宝",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "remark",
            "in": "query",
            "description": "备注",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "下单用户",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userName",
            "in": "query",
            "description": "用户名",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«Page«订单表»»",
              "originalRef": "返回值«Page«订单表»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "tags": [
          "订单"
        ],
        "summary": "查询",
        "operationId": "getUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«订单表»",
              "originalRef": "返回值«订单表»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean"
        },
        "column": {
          "type": "string"
        }
      },
      "title": "OrderItem"
    },
...
    "Page«订单表»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "hitCount": {
          "type": "boolean"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/订单表",
            "originalRef": "订单表"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«订单表»"
    },
...
    "订单明细表": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number",
          "description": "金额"
        },
        "dishFlavor": {
          "type": "string",
          "description": "口味"
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品id"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "image": {
          "type": "string",
          "description": "图片"
        },
        "name": {
          "type": "string",
          "description": "名称"
        },
        "number": {
          "type": "integer",
          "format": "int32",
          "description": "数量"
        },
        "orderId": {
          "type": "integer",
          "format": "int64",
          "description": "订单id"
        },
        "setmealId": {
          "type": "integer",
          "format": "int64",
          "description": "套餐id"
        }
      },
      "title": "订单明细表"
    },
    "订单表": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "description": "地址"
        },
        "addressBookId": {
          "type": "integer",
          "format": "int64",
          "description": "地址"
        },
        "amount": {
          "type": "number",
          "description": "实收金额"
        },
        "checkoutTime": {
          "type": "string",
          "format": "date-time",
          "description": "结账时间"
        },
        "consignee": {
          "type": "string",
          "description": "收货人"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "number": {
          "type": "string",
          "description": "订单号"
        },
        "orderDetails": {
          "type": "array",
          "description": "明细",
          "items": {
            "$ref": "#/definitions/订单明细表",
            "originalRef": "订单明细表"
          }
        },
        "orderTime": {
          "type": "string",
          "format": "date-time",
          "description": "下单时间"
        },
        "payMethod": {
          "type": "integer",
          "format": "int32",
          "description": "支付方式 1微信，2支付宝"
        },
        "phone": {
          "type": "string",
          "description": "手机号"
        },
        "remark": {
          "type": "string",
          "description": "备注"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消"
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "下单用户"
        },
        "userName": {
          "type": "string",
          "description": "用户名"
        }
      },
      "title": "订单表"
    },
    "订单表0": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "description": "地址"
        },
        "addressBookId": {
          "type": "integer",
          "format": "int64",
          "description": "地址"
        },
        "amount": {
          "type": "number",
          "description": "实收金额"
        },
        "checkoutTime": {
          "type": "string",
          "format": "date-time",
          "description": "结账时间"
        },
        "consignee": {
          "type": "string",
          "description": "收货人"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "number": {
          "type": "string",
          "description": "订单号"
        },
        "orderTime": {
          "type": "string",
          "format": "date-time",
          "description": "下单时间"
        },
        "payMethod": {
          "type": "integer",
          "format": "int32",
          "description": "支付方式 1微信，2支付宝"
        },
        "phone": {
          "type": "string",
          "description": "手机号"
        },
        "remark": {
          "type": "string",
          "description": "备注"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消"
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "下单用户"
        },
        "userName": {
          "type": "string",
          "description": "用户名"
        }
      },
      "title": "订单表0"
    },
    "订单表1": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "description": "地址"
        },
        "addressBookId": {
          "type": "integer",
          "format": "int64",
          "description": "地址"
        },
        "amount": {
          "type": "number",
          "description": "实收金额"
        },
        "checkoutTime": {
          "type": "string",
          "format": "date-time",
          "description": "结账时间"
        },
        "consignee": {
          "type": "string",
          "description": "收货人"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "number": {
          "type": "string",
          "description": "订单号"
        },
        "orderTime": {
          "type": "string",
          "format": "date-time",
          "description": "下单时间"
        },
        "payMethod": {
          "type": "integer",
          "format": "int32",
          "description": "支付方式 1微信，2支付宝"
        },
        "phone": {
          "type": "string",
          "description": "手机号"
        },
        "remark": {
          "type": "string",
          "description": "备注"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消"
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "下单用户"
        },
        "userName": {
          "type": "string",
          "description": "用户名"
        }
      },
      "title": "订单表1"
    },
...
    "返回值": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "object",
          "description": "返回对象"
        },
        "map": {
          "type": "object",
          "description": "动态字段"
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值"
    },
...
    "返回值«Page«订单表»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/Page«订单表»",
          "originalRef": "Page«订单表»"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«Page«订单表»»"
    },
...
    "返回值«订单表»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/订单表0",
          "originalRef": "订单表0"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«订单表»"
    },
...
  }
}